% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcps.R
\docType{package}
\name{bcps}
\alias{bcps}
\title{BCPS: A package with data on Baltimore City Public Schools.}
\description{
The bcps package includes demographic, enrollment, and survey data related to
Baltimore City Public Schools. Additional data from the Maryland State Department
of Education and other sources will be added soon.
}
