% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bcps_es_zones_SY2122}
\alias{bcps_es_zones_SY2122}
\title{Baltimore City Public School Elementary School Attendance Zones (SY 2021-2022)}
\format{
A data frame with 92 features and 5 variables:
\describe{
\item{\code{zone_number}}{Program zone number}
\item{\code{zone_name}}{Program zone name}
\item{\code{program_number}}{Program number (school number)}
\item{\code{program_name_short}}{Program name (short)}
\item{\code{geometry}}{Multipolygon geometry for zone boundaries}
}
}
\source{
\url{https://services3.arcgis.com/mbYrzb5fKcXcAMNi/ArcGIS/rest/services/BCPSZones_2122/FeatureServer/0}
}
\usage{
bcps_es_zones_SY2122
}
\description{
Boundaries for Baltimore City Public School elementary school attendance zones for
the 2021-2022 school year. The \code{program_number} variable is equivalent to the \code{school_number}
variable used in the demographic and survey datasets.
}
\keyword{datasets}
