% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bcps_es_zones_SY2223}
\alias{bcps_es_zones_SY2223}
\title{Baltimore City Public School Elementary School Attendance Zones (SY
2022-2023)}
\format{
A data frame with 92 rows and 9 variables:
\describe{
\item{\code{zone_number}}{Program zone number}
\item{\code{zone_name}}{Program zone name}
\item{\code{program_number}}{Program number (school number)}
\item{\code{program_name_short}}{Program name (short)}
\item{\code{ms_feeder}}{Flag for elementary school programs that feed into a middle school}
\item{\code{lower_grades}}{Description for lower grades}
\item{\code{upper_grades}}{Description for higher grades (identifying middle school that the elementary feeds into if appropriate)}
\item{\code{quadrant}}{Quadrant of Baltimore City}
\item{\code{geometry}}{Multipolygon geometry for zone boundaries}
}
}
\source{
\url{https://services3.arcgis.com/mbYrzb5fKcXcAMNi/ArcGIS/rest/services/SY2223_Elementary_Zones/FeatureServer/21}
}
\usage{
bcps_es_zones_SY2223
}
\description{
Boundaries for Baltimore City Public School elementary school attendance
zones for the 2022-2023 school year. The \code{program_number} variable is
equivalent to the \code{school_number} variable used in the demographic and survey
datasets.
}
\keyword{datasets}
