% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bcps_es_zones_SY2324}
\alias{bcps_es_zones_SY2324}
\title{Baltimore City Public School Elementary School Attendance Zones (SY
2023-2024)}
\format{
A data frame with 90 rows and 6 variables:
\describe{
\item{\code{zone_number}}{Program zone number}
\item{\code{zone_name}}{Program zone name}
\item{\code{program_number}}{Program number (school number)}
\item{\code{program_name}}{Program name}
\item{\code{program_name_short}}{Program name (short)}
\item{\code{upper_grades}}{Description for higher grades (identifying middle school that the elementary feeds into if appropriate)}
\item{\code{geometry}}{Multipolygon geometry for zone boundaries}
}
}
\source{
\url{https://services3.arcgis.com/mbYrzb5fKcXcAMNi/arcgis/rest/services/School_Profiles_WFL1/FeatureServer/14}
}
\usage{
bcps_es_zones_SY2324
}
\description{
Boundaries for Baltimore City Public School elementary school attendance
zones for the 2023-2024 school year. The \code{program_number} variable is
equivalent to the \code{school_number} variable used in the demographic and survey
datasets.
}
\keyword{datasets}
