% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bcps_programs_SY1920}
\alias{bcps_programs_SY1920}
\title{Baltimore City Public School Program Locations (SY 2019-2020)}
\format{
A data frame with 174 rows and 11 variables:
\describe{
\item{\code{program_number}}{Program number (school number)}
\item{\code{program_name}}{Program name}
\item{\code{program_name_short}}{Program name (short)}
\item{\code{grade_band}}{Grade band for program}
\item{\code{management_type}}{School management type}
\item{\code{swing}}{Logical indicator for program at swing site}
\item{\code{swing_building_number}}{Building number for swing site}
\item{\code{colocated}}{Logical indicator for programs colocated with another
program}
\item{\code{council_district}}{Baltimore City Council District Number}
\item{\code{legislative_district}}{Maryland Legislative District}
\item{\code{geometry}}{Point geometry for program/school locations}
}
}
\source{
\url{https://services3.arcgis.com/mbYrzb5fKcXcAMNi/ArcGIS/rest/services/SY1920_Building_Status_Publish/FeatureServer/0}
}
\usage{
bcps_programs_SY1920
}
\description{
Locations for Baltimore City Public School buildings for the 2019-2020
school year. The \code{program_number} variable is equivalent to the \code{school_number}
variable used in the demographic and survey datasets.
}
\keyword{datasets}
