% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bcps_programs_SY2021}
\alias{bcps_programs_SY2021}
\title{Baltimore City Public School Program Locations (SY 2020-2021)}
\format{
A data frame with 165 rows and 18 variables:
\describe{
  \item{program_number}{Program number (school number)}
  \item{program_name_short}{Program name (short)}
  \item{grade_band}{Grade band for program}
  \item{management_type}{School management type}
  \item{category}{Grade category}
  \item{swing}{Logical indicator for program at swing site}
  \item{swing_building_number}{Building number for swing site}
  \item{colocated}{Logical indicator for programs colocated with another program}
  \item{two_buildings}{Logical indicator for programs with two buildings at one site}
  \item{home_building}{...}
  \item{council_district}{Baltimore City Council District Number}
  \item{legislative_district}{Maryland Legislative District}
  \item{congressional_district}{U.S. Congressional District}
  \item{planning_area}{Planning area, Baltimore City Department of Planning}
  \item{quadrant}{Quadrant}
  \item{zone_number}{Program zone number}
  \item{zone_name}{Program zone name}
  \item{geometry}{Multipolygon geometry for zone boundaries}
}
}
\source{
\url{https://services3.arcgis.com/mbYrzb5fKcXcAMNi/ArcGIS/rest/services/SY2021_Programs/FeatureServer}
}
\usage{
bcps_programs_SY2021
}
\description{
Locations for Baltimore City Public School buildings and programs for the 2020-2021
school year. The \code{program_name} variable is equivalent to the \code{school_number}
variable used in the demographic and survey datasets.
}
\keyword{datasets}
