% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bcps_programs_SY2223}
\alias{bcps_programs_SY2223}
\title{Baltimore City Public School Program Locations (SY 2022-2023)}
\format{
A data frame with 159 rows and 8 variables:
\describe{
\item{\code{program_number}}{Program number (school number)}
\item{\code{program_name_short}}{Program name (short)}
\item{\code{grade_band}}{Grade band for program}
\item{\code{management_type}}{School management type}
\item{\code{category}}{Grade category}
\item{\code{swing}}{Logical indicator for program at swing site}
\item{\code{swing_building_number}}{Building number for swing site}
\item{\code{geometry}}{Point geometry for program/school locations}
}
}
\source{
\url{https://services3.arcgis.com/mbYrzb5fKcXcAMNi/ArcGIS/rest/services/SY2223_Programs/FeatureServer/0}
}
\usage{
bcps_programs_SY2223
}
\description{
Locations for Baltimore City Public School buildings and programs for the
2022-2023 school year. The \code{program_number} variable is equivalent to the
\code{school_number} variable used in the demographic and survey datasets.
}
\keyword{datasets}
