% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bcps_surplus_schools}
\alias{bcps_surplus_schools}
\title{Baltimore City Public Schools Surplus School Buildings}
\format{
A data frame with 40 rows and 15 variables:
\describe{
\item{\code{pos}}{Position/sort order}
\item{\code{id}}{Identifier}
\item{\code{building_name}}{Building name}
\item{\code{primary_building_year}}{Primary building year of construction}
\item{\code{address}}{Building street address}
\item{\code{school_name}}{School name}
\item{\code{school_number}}{School number}
\item{\code{school_closure_year}}{Year school closed}
\item{\code{surplus_year}}{Year surplussed}
\item{\code{status_sy20}}{Status (Fall 2020)}
\item{\code{status_sy22}}{Status (Fall 2022)}
\item{\code{building_sq_ft}}{Building interior area (square feet)}
\item{\code{property_acres}}{Property area (acres)}
\item{\code{flag_property_acres}}{Flag for properties with shared campus and
inaccurate acreage}
\item{\code{url}}{URL for Baltimore City Surplus Schools website or related
ArcGIS Feature Layer}
\item{\code{geometry}}{POINT geometry with building location}
}
}
\source{
\href{https://services1.arcgis.com/43Lm3JYE3nM91DAF/ArcGIS/rest/services/Surplus_Schools/FeatureServer/0}{Baltimore City Department of Planning - FeatureLayer}
}
\usage{
bcps_surplus_schools
}
\description{
Baltimore City Public Schools Surplus School Buildings
}
\keyword{datasets}
