% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcpss-package.R
\docType{package}
\name{bcpss-package}
\alias{bcpss}
\alias{bcpss-package}
\title{bcpss: Baltimore City Public Schools}
\description{
Provides access to demographic, enrollment, and survey data on Baltimore City Public School System (BCPSS).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/elipousson/bcpss}
  \item \url{https://elipousson.github.io/bcpss/}
  \item Report bugs at \url{https://github.com/elipousson/bcpss/issues}
}

}
\author{
\strong{Maintainer}: Eli Pousson \email{eli.pousson@gmail.com} (\href{https://orcid.org/0000-0001-8280-1706}{ORCID})

}
\keyword{internal}
