% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcpss.R
\docType{package}
\name{bcpss}
\alias{bcpss}
\title{BCPSS: A package with data on Baltimore City Public School System.}
\description{
The bcpss package includes demographic, enrollment, and survey data related to the
Baltimore City Public School System (BCPSS). Additional data from the Maryland State Department
of Education and other sources will be added soon.
}
