% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bcpss_buildings_SY2122}
\alias{bcpss_buildings_SY2122}
\title{BCPSS Buildings (SY 2021-2022) from CEFMP}
\format{
A data frame with 173 rows and 18 variables:
\describe{
\item{\code{psc_number}}{PSC Number}
\item{\code{bldg_number}}{Building number}
\item{\code{bldg_name}}{Building name}
\item{\code{program_number}}{Program number}
\item{\code{program_name}}{Program name}
\item{\code{management_type}}{Management type}
\item{\code{address}}{Address}
\item{\code{src}}{SRC (State Rated Capacity)}
\item{\code{bldg_enrollment}}{Total building enrollment}
\item{\code{program_enrollment}}{Program enrollment}
\item{\code{utilization_pct}}{Utilitization (\% of SRC)}
\item{\code{acres}}{Acres}
\item{\code{bldg_sqft}}{Building sq. feet}
\item{\code{fci_score}}{FCI Score (physical condition)}
\item{\code{notes}}{Building notes}
\item{\code{swing_space}}{Swing space indicator}
\item{\code{bcpss_owned}}{BCPSS owned building indicator}
\item{\code{year_built}}{Year built}
}
}
\usage{
bcpss_buildings_SY2122
}
\description{
Buildings and programs within BCPSS. Note that each row is a program (not a
building).
}
\keyword{datasets}
