% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bcpss_student_mobility}
\alias{bcpss_student_mobility}
\title{BCPSS Student Mobility (SY 2003-2022)}
\format{
A data frame with 5586 rows and 14 variables:
\describe{
\item{\code{year}}{School or academic year for enrollment count, e.g.
2019 data is from the start of the 2019-2020 school year.}
\item{\code{school_number}}{School number as integer (0 indicates all
schools)}
\item{\code{school_name}}{School name}
\item{\code{school_type}}{School type (e.g. Elementary, Middle, High)}
\item{\code{mobile_pct}}{Student total mobility rate}
\item{\code{entry_pct}}{Student entry mobility rate}
\item{\code{exit_pct}}{Student exit mobility rate}
\item{\code{mobile_count}}{Sudent entry and exit count}
\item{\code{entry_count}}{Student entry count}
\item{\code{exit_count}}{Student exit count}
\item{\code{avg_enrolled_count}}{Average daily student enrollment.}
\item{\code{date_created}}{Date record created.}
\item{\code{lss_number}}{Local school system (LSS) number as integer. NA is used
in place of a LSS number for statewide data. Older LEA numbers are combined
with LSS numbers in this dataset.}
\item{\code{lss_name}}{Local school system (LSS) names (typically same as county
name). "State" is used in place of a LSS name for statewide data. Older LEA
names are combined with LSS names in this dataset.}
}
}
\source{
\href{https://reportcard.msde.maryland.gov/Graphs/#/DataDownloads/datadownload}{MSDE Data Downloads}
}
\usage{
bcpss_student_mobility
}
\description{
Student mobility data from the Maryland State Department of Education (MSDE).
A subset of the \link[marylandedu:msde_student_mobility]{marylandedu::msde_student_mobility} dataset.
}
\keyword{datasets}
