% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{enrollment_demographics_SY1920}
\alias{enrollment_demographics_SY1920}
\title{Enrollment and demographic data by Baltimore City Public School (SY 2019-2020)}
\format{

}
\source{
\url{https://www.baltimorecityschools.org/sites/default/files/inline-files/SY19-20-Enrollment-11.10.2020.xlsx}
}
\usage{
enrollment_demographics_SY1920
}
\description{
Total enrollment and demographic characteristics for Baltimore City
public school students including data aggregated by grade, grade band,
school, and citywide for the 2019-2020 school year.
}
\keyword{datasets}
