% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{enrollment_demographics_SY1920_long}
\alias{enrollment_demographics_SY1920_long}
\title{Baltimore City Public Schools Enrollment and Demographic data (SY 2019-2020)
\itemize{
\item long format
}}
\format{
A data frame with 26,916 rows and 10 variables:
\describe{
\item{\code{school_number}}{School number}
\item{\code{school_name}}{School name}
\item{\code{management_type}}{School management type}
\item{\code{grade_band}}{Grade band}
\item{\code{grade}}{Grade}
\item{\code{grade_range}}{Grade range}
\item{\code{total_enrollment}}{Total Enrollment}
\item{\code{variable}}{Variable}
\item{\code{share}}{Percent of Total Enrollment}
\item{\code{label}}{Variable label}
}
}
\source{
\url{https://www.baltimorecityschools.org/data}
}
\usage{
enrollment_demographics_SY1920_long
}
\description{
Total enrollment and demographic characteristics for Baltimore City public
school students including data aggregated by grade, grade band, school, and
citywide for the 2019-2020 school year. Long format version is based on the
\code{enrollment_demographics_SY1920} data with an additional \code{label} column based
on the original Excel sheet column names.
}
\keyword{datasets}
