% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{enrollment_demographics_SY2122}
\alias{enrollment_demographics_SY2122}
\title{Baltimore City Public Schools Enrollment and Demographic data (SY 2021-2022)}
\format{
A data frame with 2,153 rows and 20 variables:
\describe{
\item{\code{school_number}}{School number}
\item{\code{school_name}}{School name}
\item{\code{management_type}}{School management type}
\item{\code{grade_band}}{Grade band (factor)}
\item{\code{grade}}{Grade (factor)}
\item{\code{grade_range}}{Grade range (factor)}
\item{\code{total_enrollment}}{Total Enrollment}
\item{\code{percent_male}}{Percent Male}
\item{\code{percent_female}}{Percent Female}
\item{\code{percent_nonbinary}}{Percent Nonbinary}
\item{\code{percent_direct_certification}}{Percent Direct Certification}
\item{\code{percent_swd}}{Percent Students with Disabilities (SWD)}
\item{\code{percent_el}}{Percent English Learner (EL)}
\item{\code{percent_non_hispanic_black}}{Percent Non-Hispanic Black}
\item{\code{percent_non_hispanic_white}}{Percent Non-Hispanic White}
\item{\code{percent_hispanic}}{Percent Hispanic}
\item{\code{percent_non_hispanic_asian}}{Percent Non-Hispanic Asian}
\item{\code{percent_non_hispanic_american_indian}}{Percent Non-Hispanic American Indian}
\item{\code{percent_non_hispanic_pacific_islander}}{Percent Non-Hispanic Native Hawaiian or Other Pacific Islander}
\item{\code{percent_non_hispanic_multiracial}}{Percent Non-Hispanic Multiracial}
}
}
\source{
\url{https://www.baltimorecityschools.org/data}
}
\usage{
enrollment_demographics_SY2122
}
\description{
Total enrollment and demographic characteristics for Baltimore City public
school students including data aggregated by grade, grade band, school, and
citywide for the 2021-2022 school year.
}
\details{
The column \code{percent_non_hispanic_pacific_islander} is named Percent
non-Hispanic Hawaiian and Other Pacific Islander in the source but is renamed
to match the prior two years of data. This data also includes a
`percent_nonbinary column that is not present in prior years.

For variables with percentages, a value of 0.05 may represent <=5\% rather
than 5\% and a value of 0.95 may represent >=95\% rather than 95\%.
}
\keyword{datasets}
