% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{enrollment_msde_SY0919}
\alias{enrollment_msde_SY0919}
\title{Enrolled students by school and grade (SY 2009-2019)}
\format{
A data frame with 16,763 rows and 6 variables:
\describe{
 \item{\code{school_year}}{School or academic year for enrollment count, e.g.
   2019 data is from the start of the 2019-2020 school year.}
 \item{\code{school_number}}{School number (0 indicates all schools)}
 \item{\code{school_name}}{School name}
 \item{\code{enrolled_count}}{Count of enrolled students in the grade or grade range}
 \item{\code{grade}}{Grades from Prekindergarden (PK) to Grade 12}
 \item{\code{grade_range}}{Elementary school, middle school, high school, or
   all grades (total enrollment)}
 }
}
\source{
\url{https://reportcard.msde.maryland.gov/Graphs/#/DataDownloads//3/17/6/99/XXXX/2019}
}
\usage{
enrollment_msde_SY0919
}
\description{
Enrollment data from the Maryland State Department of Education.
 Raw CSV data files with the full state enrollment data are available in the
 \code{inst/extdata} folder.
}
\keyword{datasets}
