% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fmt_school_name_short}
\alias{fmt_school_name_short}
\alias{vec_fmt_school_name_short}
\title{Shorten a column or character vector of school names}
\usage{
fmt_school_name_short(
  data,
  .col = "school_name",
  .col_short = NULL,
  suffix = "S",
  extra_pattern = c(`All Baltimore City Public Schools` = "All City schools", Academy =
    "Acad.", Preparatory = "Prep.", Square = "Sq.", `^Commodore` = "Com.", `, The$` = "")
)

vec_fmt_school_name_short(
  string,
  suffix = "S",
  extra_pattern = c(`All Baltimore City Public Schools` = "All City schools", Academy =
    "Acad.", Preparatory = "Prep.", Square = "Sq.", `^Commodore` = "Com.", `, The$` = "")
)
}
\arguments{
\item{data}{Data frame with column name matching .col that contains school
names to shorten.}

\item{.col}{Column name to shorten. Values are replaced if .col_short is
supplied.}

\item{.col_short}{Column name for new column with short names.}

\item{suffix}{Suffix to append to standard abbreviations. Defaults to "S"}

\item{extra_pattern}{Additional replacements added to standard pattern and
passed to \code{\link[stringr:str_replace]{stringr::str_replace_all()}}}

\item{string}{Character vector of school names to shorten.}
}
\description{
Standard grade range names are replaced with corresponding abbreviations
("E","EM","M","H","MH", and "P"). For example, "High" (appearing at the end
of a string), "High school" or "High School" is replaced by HS (or H if
suffix is \code{NULL}). String or column values are also passed to
\code{\link[stringr:str_trim]{stringr::str_squish()}} and \code{\link[stringr:str_trim]{stringr::str_trim()}} to avoid issues with
pattern matching failing due to trailing spaces or double-spaces.
}
\keyword{internal}
