% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nces_school_directory_SY1920}
\alias{nces_school_directory_SY1920}
\title{National Center for Education Statistics (NCES) School Directory (SY 2019-2020)}
\format{
A data frame with 167 rows and 10 variables:
\describe{
\item{\code{school_year}}{School year}
\item{\code{school_number}}{School number}
\item{\code{school_name}}{School name}
\item{\code{grade_band}}{Grade band (school type)}
\item{\code{address}}{Street address}
\item{\code{city}}{City}
\item{\code{state}}{State}
\item{\code{zip}}{Zipcode}
\item{\code{phone}}{Phone number}
\item{\code{nces_number}}{NCES number}
}
}
\source{
\href{https://reportcard.msde.maryland.gov/Graphs/#/DataDownloads/datadownload}{MSDE Data Downloads}
}
\usage{
nces_school_directory_SY1920
}
\description{
NCES Directory listing for Baltimore City Schools including school addresses,
phone numbers, and NCES numbers.
}
\keyword{datasets}
