% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{parent_school_survey_SY1819_long}
\alias{parent_school_survey_SY1819_long}
\title{Baltimore City Public School Parent Survey Results (2018-2019 school year) - long format}
\format{
A data frame with 7,224 rows and 10 variables:
\describe{
  \item{school_number}{School Number}
  \item{school_name}{School Name}
  \item{school_type}{School Type}
  \item{grade_band}{Grade Band}
  \item{parent_response_count}{Number of parent survey responses}
  \item{estimated_parent_number_eligible}{Estimated number of eligible parents}
  \item{estimated_parent_response_rate}{Estimated parent response rate}
  \item{variable}{Unique slug for composite score or question}
  \item{value}{Value of composite score or share of parents responding affirmatively to question}
  \item{label}{Variable label for composite score or question}
}
}
\source{

}
\usage{
parent_school_survey_SY1819_long
}
\description{
Results from the Baltimore City Public School Parent Survey administered
during the for the 2018-2029 school year. Data includes composite scores
and responses to selected questions school. Long format version is
based on the `parent_school_survey_SY1819` data with an additional `label`
column based on the original Excel sheet column names.
}
\keyword{datasets}
