% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{parent_survey_SY1819_long}
\alias{parent_survey_SY1819_long}
\title{Baltimore City Public School Parent Survey Results (SY 2018-2019) -
long format}
\format{
A data frame with 7,224 rows and 10 variables:
\describe{
  \item{\code{school_number}}{School number}
  \item{\code{school_name}}{School name}
  \item{\code{management_type}}{School management type}
  \item{\code{grade_band}}{School grade band}
  \item{\code{parent_response_count}}{Number of parent survey responses}
  \item{\code{est_num_parents}}{Estimated number of eligible parents}
  \item{\code{est_response_rate}}{Estimated parent response rate}
  \item{\code{variable}}{Unique slug for composite score or question}
  \item{\code{value}}{Value of composite score or share of parents responding affirmatively to question}
  \item{\code{label}}{Variable label for composite score or question}
}
}
\source{
\url{https://www.baltimorecityschools.org/data}
}
\usage{
parent_survey_SY1819_long
}
\description{
Results from the Baltimore City Public School Parent Survey administered
during the for the 2018-2029 school year. Data includes composite scores and
responses to selected questions school. Long format version is based on the
\code{parent_survey_SY1819} data with an additional \code{label} column based on the
original Excel sheet column names.
}
\keyword{datasets}
