% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{attendance_msde_SY0919}
\alias{attendance_msde_SY0919}
\title{Attendance data by school (SY 2009-2019)}
\format{
A data frame with 3,197 rows and 15 variables:
\describe{
  \item{\code{school_year}}{School year or academic year, e.g. 2019 for 2019-2020 school year}
  \item{\code{school_number}}{School number (0 indicates all schools)}
  \item{\code{school_name}}{School name}
  \item{\code{grade_band}}{Grade band (school type)}
  \item{\code{attend_rate_pct}}{Attendance rate (percent)}
  \item{\code{days_attended_cnt}}{Count of days attended}
  \item{\code{days_member_cnt}}{double ... count}
  \item{\code{fewer_5_pct}}{double ... percent}
  \item{\code{fewer_5_cnt}}{double ... count}
  \item{\code{more_20_pct}}{double ... percent}
  \item{\code{more_20_cnt}}{double ... count}
  \item{\code{more_90_member_cnt}}{double ... count}
  \item{\code{chronic_absentee_cnt}}{Chronic absentee count}
  \item{\code{chronic_absentee_denom}}{Chronic absentee denominator}
  \item{\code{chronic_absentee_pct}}{Chronic absentee percent}
}
}
\source{
\url{https://reportcard.msde.maryland.gov/Graphs/#/DataDownloads//3/17/6/99/XXXX/2019}
}
\usage{
attendance_msde_SY0919
}
\description{
Attendance data from the Maryland State Department of Education.
  Raw CSV data files with the full state enrollment data are available in the
  \code{inst/extdata} folder. Any percent variables with a value of 95
  percent should be interpreted as equivalent at or above 95 percent. Any
  percent variables with a value of 5 percent should be interpreted as
  equivalent at or below 5 percent. Chronic absentee data is not available
  prior to 2018. Variable definitions are incomplete.
}
\keyword{datasets}
