% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bcps_es_zones_SY2021}
\alias{bcps_es_zones_SY2021}
\title{Baltimore City Public School Elementary School Attendance Zones (SY 2020-2021)}
\format{
A data frame with 96 rows and 5 variables:
\describe{
  \item{zone_number}{Program zone number}
  \item{zone_name}{Program zone name}
  \item{program_number}{Program number (school number)}
  \item{program_name_short}{Program name (short)}
  \item{geometry}{Multipolygon geometry for zone boundaries}
}
}
\source{
\url{https://services3.arcgis.com/mbYrzb5fKcXcAMNi/ArcGIS/rest/services/BCPSZones_2021/FeatureServer/0}
}
\usage{
bcps_es_zones_SY2021
}
\description{
Boundaries for Baltimore City Public School elementary school attendance zones for
the 2020-2021 school year. The `program_name` variable is equivalent to the `school_number`
variable used in the demographic and survey datasets.
}
\keyword{datasets}
