% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bcpss_attendance}
\alias{bcpss_attendance}
\title{BCPSS Attendance Data}
\format{
A data frame with 5585 rows and 18 variables:
\describe{
\item{\code{year}}{integer COLUMN_DESCRIPTION}
\item{\code{school_number}}{integer COLUMN_DESCRIPTION}
\item{\code{school_name}}{character COLUMN_DESCRIPTION}
\item{\code{school_type}}{character COLUMN_DESCRIPTION}
\item{\code{attend_rate_pct}}{double COLUMN_DESCRIPTION}
\item{\code{days_attended_count}}{double COLUMN_DESCRIPTION}
\item{\code{days_member_count}}{double COLUMN_DESCRIPTION}
\item{\code{fewer_5_pct}}{double COLUMN_DESCRIPTION}
\item{\code{fewer_5_count}}{double COLUMN_DESCRIPTION}
\item{\code{more_20_pct}}{double COLUMN_DESCRIPTION}
\item{\code{more_20_count}}{double COLUMN_DESCRIPTION}
\item{\code{more_90_member_count}}{double COLUMN_DESCRIPTION}
\item{\code{chronic_absentee_count}}{double COLUMN_DESCRIPTION}
\item{\code{chronic_absentee_denom}}{double COLUMN_DESCRIPTION}
\item{\code{chronic_absentee_pct}}{double COLUMN_DESCRIPTION}
\item{\code{date_created}}{double COLUMN_DESCRIPTION}
\item{\code{lss_name}}{character COLUMN_DESCRIPTION}
\item{\code{lss_number}}{integer COLUMN_DESCRIPTION}
}
}
\usage{
bcpss_attendance
}
\description{
This documentation is a place-holder.
}
\details{
DETAILS
}
\keyword{datasets}
