% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bcpss_enrollment}
\alias{bcpss_enrollment}
\title{BCPSS Enrollment Data}
\format{
A data frame with 31618 rows and 10 variables:
\describe{
\item{\code{year}}{integer COLUMN_DESCRIPTION}
\item{\code{school_number}}{integer COLUMN_DESCRIPTION}
\item{\code{school_name}}{character COLUMN_DESCRIPTION}
\item{\code{enrolled_count}}{integer COLUMN_DESCRIPTION}
\item{\code{grade}}{character COLUMN_DESCRIPTION}
\item{\code{grade_range}}{character COLUMN_DESCRIPTION}
\item{\code{race}}{character COLUMN_DESCRIPTION}
\item{\code{date_created}}{double COLUMN_DESCRIPTION}
\item{\code{lss_number}}{integer COLUMN_DESCRIPTION}
\item{\code{lss_name}}{character COLUMN_DESCRIPTION}
}
}
\usage{
bcpss_enrollment
}
\description{
This documentation is a place-holder.
}
\details{
DETAILS
}
\keyword{datasets}
