% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bcpss_student_mobility}
\alias{bcpss_student_mobility}
\title{BCPSS Student Mobility Data}
\format{
A data frame with 5586 rows and 14 variables:
\describe{
\item{\code{year}}{integer COLUMN_DESCRIPTION}
\item{\code{school_number}}{integer COLUMN_DESCRIPTION}
\item{\code{school_name}}{character COLUMN_DESCRIPTION}
\item{\code{school_type}}{character COLUMN_DESCRIPTION}
\item{\code{mobile_pct}}{double COLUMN_DESCRIPTION}
\item{\code{entry_pct}}{double COLUMN_DESCRIPTION}
\item{\code{exit_pct}}{double COLUMN_DESCRIPTION}
\item{\code{mobile_count}}{integer COLUMN_DESCRIPTION}
\item{\code{entry_count}}{integer COLUMN_DESCRIPTION}
\item{\code{exit_count}}{integer COLUMN_DESCRIPTION}
\item{\code{avg_enrolled_count}}{double COLUMN_DESCRIPTION}
\item{\code{lss_name}}{character COLUMN_DESCRIPTION}
\item{\code{date_created}}{double COLUMN_DESCRIPTION}
\item{\code{lss_number}}{integer COLUMN_DESCRIPTION}
}
}
\usage{
bcpss_student_mobility
}
\description{
This documentation is a place-holder.
}
\details{
DETAILS
}
\keyword{datasets}
