% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{enrollment_demographics_SY1920_long}
\alias{enrollment_demographics_SY1920_long}
\title{Enrollment and demographic data by Baltimore City Public School (SY 2019-2020) - long format}
\format{
A data frame with 26,916 rows and 10 variables:
\describe{
  \item{school_number}{School number}
  \item{school_name}{School name}
  \item{management_type}{School management type}
  \item{grade_band}{Grade band}
  \item{grade}{Grade}
  \item{grade_range}{Grade range}
  \item{total_enrollment}{Total Enrollment}
  \item{variable}{Variable}
  \item{share}{Percent of Total Enrollment}
  \item{label}{Variable label}
}
}
\source{
\url{https://www.baltimorecityschools.org/data}
}
\usage{
enrollment_demographics_SY1920_long
}
\description{
Total enrollment and demographic characteristics for Baltimore City
public school students including data aggregated by grade, grade band,
school, and citywide for the 2019-2020 school year. Long format version is
based on the `enrollment_demographics_SY1920` data with an additional `label`
column based on the original Excel sheet column names.
}
\keyword{datasets}
