% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{enrollment_demographics_SY1920_long}
\alias{enrollment_demographics_SY1920_long}
\title{Enrollment and demographic data by Baltimore City Public School (SY 2019-2020) - long format}
\format{

}
\source{
\url{https://www.baltimorecityschools.org/sites/default/files/inline-files/SY19-20-Enrollment-11.10.2020.xlsx}
}
\usage{
enrollment_demographics_SY1920_long
}
\description{
Total enrollment and demographic characteristics for Baltimore City
public school students including data aggregated by grade, grade band,
school, and citywide for the 2019-2020 school year. Long format version is
based on the `enrollment_demographics_SY1920` data with an additional `label`
column based on the original Excel sheet column names.
}
\keyword{datasets}
